pyproject.toml
[project]
name = "ollama-cli"
version = "0.1.0"
description = "CLI tool to query Ollama LLM for terminal help"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.scripts]
ollama-ask = "ollama_cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ollama_cli"]